added "other household" place, initialized with a random household. Persons will go there on the other household activity.
git-svn-id: https://scrap.dis.anl.gov/repos/MRSA/models/agentmodel@264 eb7698ad-c3c5-4393-a60c-5d690fe7bf3d

+++ MRSA.HPC/src/Activity.h
@@ -70,7 +70,7 @@ public:
 
 private:
 	std::string tucase_id, place_type;
-	float start_time, end_time;
+	int start_time, end_time;
 	int activity_risk;
 	float place_risk;
